+++ packages/@expressive-code/plugin-shiki/src/index.ts
@@ -1,6 +1,6 @@
 import { ExpressiveCodePlugin, ExpressiveCodeTheme, InlineStyleAnnotation } from '@expressive-code/core'
 import { getCachedHighlighter } from './cache'
-import { BUNDLED_THEMES, loadTheme, FontStyle, IThemedToken } from 'shiki'
+import { BUNDLED_THEMES, loadTheme, FontStyle } from 'shiki'
 
 /**
  * A list of all themes bundled with Shiki.
@@ -32,40 +32,24 @@ export function pluginShiki(): ExpressiveCodePlugin {
 			performSyntaxAnalysis: async ({ codeBlock, theme }) => {
 				const highlighter = await getCachedHighlighter({ theme })
 				const codeLines = codeBlock.getLines()
-				const tokenLines = highlighter.codeToThemedTokens(codeBlock.code, codeBlock.language, theme.name)
+				let code = codeBlock.code
+
+				// If the code block uses a terminal language and includes placeholder strings
+				// in angle brackets (e.g. `<username>`), Shiki will treat the closing `>` as
+				// a redirect operator and highlight the character before it differently.
+				// We work around this by replacing the brackets around such placeholder strings
+				// with different characters that Shiki will not interpret as operators.
+				if (isTerminalLanguage(codeBlock.language)) {
+					code = code.replace(/<([^>]*[^>\s])>/g, 'X$1X')
+				}
+
+				// Run Shiki on the code (without explanations to improve performance)
+				const tokenLines = highlighter.codeToThemedTokens(code, codeBlock.language, theme.name, { includeExplanation: false })
 				tokenLines.forEach((line, lineIndex) => {
 					let charIndex = 0
-					line.forEach((token, tokenIndex) => {
+					line.forEach((token) => {
 						const tokenLength = token.content.length
 						const tokenEndIndex = charIndex + tokenLength
-
-						// If Shiki encounters placeholders in angle brackets (e.g. `<username>`)
-						// in shell scripts, it will treat the closing angle bracket as a redirect
-						// operator and highlight the final character of the placeholder name
-						// differently. We detect this and extend the previous token instead.
-						const lastTokenWasStringUnquotedArgument =
-							tokenIndex > 0 && line[tokenIndex - 1]?.explanation?.slice(-1)[0]?.scopes?.some((scope) => scope.scopeName?.startsWith('string.unquoted.argument.'))
-						if (lastTokenWasStringUnquotedArgument) {
-							// Check if the current token is a single non-empty character,
-							// and the next token is a redirect operator
-							const isCharBeforeRedirectOperator = tokenLength === 1 && token.content.trim().length === 1 && containsRedirectOperator(line[tokenIndex + 1])
-							// Check if the current token is a two-character combination
-							// ending with a redirect operator
-							const isCharAndRedirectOperator = tokenLength === 2 && token.content.trim().length === 2 && token.content[1] === '>' && containsRedirectOperator(token)
-							if (isCharBeforeRedirectOperator || isCharAndRedirectOperator) {
-								// Move the current index forward by 1 character
-								// and extend the previous annotation to include it
-								charIndex++
-								const lastAnnotation = codeLines[lineIndex].getAnnotations().slice(-1)[0]
-								if (lastAnnotation && lastAnnotation.inlineRange) {
-									lastAnnotation.inlineRange.columnEnd = charIndex
-								}
-								// Skip processing the current token if there is nothing left
-								if (charIndex === tokenEndIndex) return
-							}
-						}
-
-						// Otherwise, add an annotation for the current token
 						const fontStyle = token.fontStyle || FontStyle.None
 						codeLines[lineIndex].addAnnotation(
 							new InlineStyleAnnotation({
@@ -87,12 +71,6 @@ export function pluginShiki(): ExpressiveCodePlugin {
 	}
 }
 
-function containsRedirectOperator(token?: IThemedToken) {
-	return token?.explanation?.some(
-		(part) =>
-			// Content must be a single ">"" character
-			part.content === '>' &&
-			// Scope name must also indicate a redirect operator
-			part.scopes?.some((scope) => scope.scopeName?.startsWith('keyword.operator.redirect.'))
-	)
+function isTerminalLanguage(language: string) {
+	return ['shellscript', 'shell', 'bash', 'sh', 'zsh'].includes(language)
 }