+++ packages/expressive-code/CHANGELOG.md
@@ -1,5 +1,13 @@
 # expressive-code
 
+## 0.4.2
+
+### Patch Changes
+
+- Turn off explanations to improve Shiki performance
+- Updated dependencies
+  - @expressive-code/plugin-shiki@0.3.1
+
 ## 0.4.1
 
 ### Patch Changes